Background renders only under cards drawn.
Previously, if set, the background color would be drawn under the whole image. This change causes the rendering methods to only draw the background under each individual card image drawn. The result is that non-full pages now don't fill unused area with the background color. This is most noticeable in the PDF rendering, where it previously would draw the entire page in the background color. Toner and ink cartridges rejoice!